+++ NEWS.md
@@ -1,3 +1,8 @@
+1.3.9
+-----
+* Do not choke on multiple `skipfields` parameter
+* Do not request caching of TRUNCATE queries
+
 1.3.8
 -----
 * Make using SET or querying system catalogues harder